feat: Add a basic CLI
The existing version has too many manual steps to run before the script can run its analysis. The new CLI is much easier to use.

Unfortunately, the old script was too hard to modify since it was tightly coupled to a bunch of global variables, so it was easier to rewrite from scratch. As a result, the pseudonyms and author overrides configs are not yet supported in the new version.