fix: Only cache `blame` results
The cache included authorship modifications from configs, but only changed when the target rev changed. As a result, changing configs for the same rev resulted in inaccurate results.

I considered two options to fix the caching:
 1. Cache only the raw `blame` authorship
 2. Incorporate configs into the cache key

Observations:
 a) `blame` is expensive to compute, but is fixed for each target revision.
 b) Authorship modifications by configs are cheap to compute AND change more frequently -- they can change even for the same target revision.

So, to improve performance, it's most helpful to cache the raw `blame` results (option 1) since it saves us the most rework, including when changing configs for the same revision (which is more frequent).

If I had instead taken option 2, hashing configs into the cache key, we'd still recompute the full blame whenever any config changes plus pay the hashing cost on every run. We'd only see a performance boost in the rare case that a rerun has the same target revision and configs.